Split the NavigationPageRenderer into partial classes for easier maintenance
